center ListItem text and rename ListItem
